fix(main): correctify wrong standalone fct calculation, use correct rtt
The standalone fct is consist of a base rtt and transmit time of file
size. The original estimation of base rtt uses the summation of
propagation delay and a transmit delay of a packet with maximum size.
However, for flow whose size is less than 1 maximum packet size, it will
bring unacceptable error: the real fct could be less than the standalone
fct which is considered to be optimal.
Thus, I remove this uncorrect transmit delay in the original code.